server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
requests==2.32.3
apscheduler==3.10.4
pytz==2024.2
